# server/controllers/GIARDINO_db/generated/Foto_fioreControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE foto_fioreControllerGenerated.js PLEASE EDIT ../foto_fioreController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import Foto_fioreModel from "../../../models/GIARDINO_db/Foto_fioreModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import Foto_fioreController from "../Foto_fioreController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/foto_fiore`;
    router.post(baseUrl + "", authorize([]), Foto_fioreController.create);
    router.delete(baseUrl + "/:id", authorize([]), Foto_fioreController.delete);
    router.get(baseUrl + "/:id", authorize([]), Foto_fioreController.get);
    router.get(baseUrl + "", authorize([]), Foto_fioreController.list);
    router.post(baseUrl + "/:id", authorize([]), Foto_fioreController.update);
  },


  // CRUD METHODS


  /**
  * foto_fioreModel.create
  *   @description CRUD ACTION create
  *   @param foto_fiore obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await Foto_fioreModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * foto_fioreModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id foto_fiore
  *
  */
  delete: async (req, res) => {
    try {
      const result = await Foto_fioreModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * foto_fioreModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id foto_fiore
  *   @returns foto_fiore
  *
  */
  get: async (req, res) => {
    try {
      const result = await Foto_fioreModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * foto_fioreModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF foto_fiore
  *
  */
  list: async (req, res) => {
    try {
      const result = await Foto_fioreModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * foto_fioreModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id foto_fiore
  *   @returns foto_fiore
  *
  */
  update: async (req, res) => {
    try {
      const result = await Foto_fioreModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;